Film Locations in San Francisco
Columns for the latitude and longitude of the film location, supervisor district, and analysis neighborhood have been added to the datasets. A blank value means geocoding was not possible based on the location description.

If you love movies, and you love San Francisco, you're bound to love this -- a listing of filming locations of movies shot in San Francisco starting from 1915. You'll find the titles, locations, fun facts, names of the director, writer, actors, and studio for most of these films.